feat: separate interfaces in other file
